Moved sensorData outside the states folder
