+++ catkin_ws/src/acr_core/src/sensorData.hpp
@@ -1,7 +1,7 @@
 #ifndef SensorData_H
 #define SensorData_H
 
-#include "../consts.hpp"
+#include "consts.hpp"
 
 #include "ros/ros.h"
 #include "sensor_msgs/Range.h"
