+++ catkin_ws/src/acr_core/src/states/state_interact.hpp
@@ -1,7 +1,7 @@
 #ifndef State_Interact_H
 #define State_Interact_H
 
-#include "sensorData.hpp"
+#include "../sensorData.hpp"
 class State_Idle;
 
 #include "ros/ros.h"
