+++ catkin_ws/src/acr_core/src/states/state_navigating.hpp
@@ -1,7 +1,7 @@
 #ifndef State_Navigating_H
 #define State_Navigating_H
 
-#include "sensorData.hpp"
+#include "../sensorData.hpp"
 class State_Idle;
 
 #include "../consts.hpp"
